Spent some time seperating files and removing some redundant code. Now each module will have a specific task, and I'll try to keep it clean
