--- convertData.py
@@ -1,1 +1,0 @@
-# fix data stored in mangaData.json to fit the new weebcentral page
